requirements/common.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0

click==8.0.1
dj-database-url==0.5.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
iniconfig==1.1.1
Markdown==3.3.4
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7

python-decouple==3.4
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
